requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
flake8>=6.1.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
streamlit>=1.30.0
pandas>=2.0.0
requests>=2.31.0
imap-tools>=1.5.0
tenacity>=8.2.3
